Observe exceptions for faulted intermediate tasks in the implementation of TaskBlocks.While
In these methods, the Finally() extension method is used to transfer the
state of one intermediate Task to another TaskCompletionSource. If the task
is faulted, only the latter needs to be observed. However, a potential
problem arises because the Task returned by Finally() *also* enters the
faulted state if the antecedent was faulted. This commit leverages the new
ObserveExceptions extension method to ensure the task returned by Finally()
does not go unobserved if the antecedent was faulted.